MET-5961 code review comments

+++ src/main/java/eu/europeana/metis/sandbox/service/debias/DetectService.java
@@ -2,25 +2,80 @@ package eu.europeana.metis.sandbox.service.debias;
 
 import eu.europeana.metis.sandbox.dto.debias.DetectionInfoDto;
 
+/**
+ * The interface Detect service.
+ */
 public interface DetectService {
 
-  void fail(Long datasetId);
+  /**
+   * Fail.
+   *
+   * @param datasetId the dataset id
+   */
+  void fail(Integer datasetId);
 
-  void success(Long datasetId);
+  /**
+   * Success.
+   *
+   * @param datasetId the dataset id
+   */
+  void success(Integer datasetId);
 
-  boolean process(Long datasetId);
+  /**
+   * Process boolean.
+   *
+   * @param datasetId the dataset id
+   * @return the boolean
+   */
+  boolean process(Integer datasetId);
 
+  /**
+   * Gets state.
+   *
+   * @return the state
+   */
   Stateful getState();
 
+  /**
+   * Sets state.
+   *
+   * @param state the state
+   */
   void setState(Stateful state);
 
+  /**
+   * Gets ready.
+   *
+   * @return the ready
+   */
   Stateful getReady();
 
+  /**
+   * Gets processing.
+   *
+   * @return the processing
+   */
   Stateful getProcessing();
 
+  /**
+   * Gets completed.
+   *
+   * @return the completed
+   */
   Stateful getCompleted();
 
+  /**
+   * Gets error.
+   *
+   * @return the error
+   */
   Stateful getError();
 
-  DetectionInfoDto getDetectionInfo(Long datasetId);
+  /**
+   * Gets detection info.
+   *
+   * @param datasetId the dataset id
+   * @return the detection info
+   */
+  DetectionInfoDto getDetectionInfo(Integer datasetId);
 }
