+++ src/main/java/eu/europeana/metis/sandbox/service/debias/Stateful.java
@@ -1,7 +1,7 @@
 package eu.europeana.metis.sandbox.service.debias;
 
 public interface Stateful {
-  void fail(Long datasetId);
-  void success(Long datasetId);
-  boolean process(Long datasetId);
+  void fail(Integer datasetId);
+  void success(Integer datasetId);
+  boolean process(Integer datasetId);
 }
